Add reconnect tests for getIntervalCollection on shared-string (#10038)
* Add reconnect tests for getIntervalCollection on shared-string
